refactor bank hour filenames
