+++ src/features/bank-hours/command.ts
@@ -1,6 +1,6 @@
 import { Command } from "../../shared/command/command";
-import { addSubcommand } from "./add-command";
-import { removeSubcommand } from "./remove-command";
+import { addSubcommand } from "./add-subcommand";
+import { removeSubcommand } from "./remove-subcommand";
 
 export const bankHourCommand = new Command(
   "bankhour",
