+++ src/features/bank-hours/remove-subcommand.ts
